% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev-fn-args.R
\name{abbrev_fn_args}
\alias{abbrev_fn_args}
\title{Enhance a function to interpret abbreviated function expressions}
\usage{
abbrev_fn_args(f, ...)
}
\arguments{
\item{f}{Function, or symbol or name of a function.}

\item{...}{Name(s) of functional argument(s) of \code{f} (strings). Unsplicing of
lists of strings is supported via \code{!!!} or \code{UQS()}.}
}
\value{
A function with the same call signature as \code{f}, but whose function
arguments, as designated by \code{...}, may also be specified as abbreviated
function expression of the form \code{.(...)}, cf. \code{\link[=as_fn]{as_fn()}}.
}
\description{
\code{abbrev_fn_args()} is a functional operator that \dQuote{upgrades} a function
to one that can interpret abbreviated function expressions.
}
\examples{
reduce <- abbrev_fn_args(Reduce, "f")
reduce(.(u, v ~ u + 1 / v), c(3, 7, 15, 1, 292), right = TRUE)

}
\seealso{
\code{\link[=as_fn]{as_fn()}}
}
