% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{\%<<<\%}
\alias{\%>>>\%}
\alias{decompose}
\title{Compose functions}
\usage{
compose(...)

f \%<<<\% g

g \%>>>\% f

decompose(f)
}
\arguments{
\item{...}{Functions to compose; lists of functions are automatically spliced
in. (Explicit \link[rlang:quasiquotation]{splicing} is supported, via \code{!!!}.)}

\item{f, g}{Functions.}
}
\value{
Composition of functions. (\code{NULL} is returned when no functions are
given.)
}
\description{
Compose functions
}
