% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{\%<<<\%}
\alias{\%>>>\%}
\alias{decompose}
\title{Compose functions}
\usage{
compose(...)

f \%<<<\% g

g \%>>>\% f

decompose(f)
}
\arguments{
\item{...}{Functions to compose; lists of functions are automatically spliced
in. (Explicit \link[rlang:quasiquotation]{splicing} is supported, via \code{!!!}.)}

\item{f, g}{Functions.}
}
\value{
\code{compose()}, \code{\%<<<\%} and \code{\%>>>\%} return a function
composition whose \link[=formals]{formals} match those of the initial function
called. \code{decompose()} returns a list of composite functions of a function
(composition).
}
\description{
Compose functions
}
