% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\alias{\%<<<\%}
\alias{\%>>>\%}
\alias{decompose}
\title{Compose functions}
\usage{
compose(...)

f \%<<<\% g

g \%>>>\% f

decompose(f)
}
\arguments{
\item{...}{Functions or lists thereof to compose. (Lists of functions are
automatically spliced in; explicit \link[rlang:quasiquotation]{splicing} via
\code{!!!} is also supported.) Following convention, functions are composed from
right to left.}

\item{f, g}{Functions.}
}
\value{
\code{compose()}, \code{\%<<<\%} and \code{\%>>>\%} return a function
composition, whose \link[=formals]{formals} match those of the initial function
called. \code{decompose()} returns the list of composite functions of a function
composition, and wraps a non-composite function in a list.
}
\description{
Compose functions in three ways:
\itemize{
\item Using \code{compose()}: \code{compose(f, g)} is the function that calls \code{g} followed
by \code{f}. It has the \link[=formals]{formals} of \code{g}.
\item Using \code{\%<<<\%} (\dQuote{backward} composition): \code{f \%<<<\% g}
is another way to express \code{compose(f, g)}.
\item Using \code{\%>>>\%} (\dQuote{forward} composition): \code{f \%>>>\% g}
is another way to express \code{compose(g, f)}.
}

Use \code{decompose()} to recover the list of composite functions of a function
composition.
}
\section{Properties}{
 \code{compose()} is \emph{associative}, semantically and
operationally. This means, for instance, that
\code{compose(f, g, h)},
\code{compose(f, compose(g, h))},
\code{compose(compose(f, g), h)},
are implemented as the \emph{same function}. In other words, lists of functions
are automatically \dQuote{flattened out} when they are composed, so there
is no danger of piling up nested compositions.

\code{decompose()} and \code{compose()} are \emph{mutually invertible}. If \code{f} is a
function, \code{compose(decompose(f))} is the same as \code{f} (regardless of
whether \code{f} is itself a composition), while if \code{fs} is a list of functions,
\code{decompose(compose(fs))} is the same as \code{fs}.
}

\examples{
# Functions are composed from right to left (according to convention)
inv <- curry(`/`)(1)  # reciprocal
f1 <- compose(inv, log, abs)
stopifnot(isTRUE(all.equal(f1(-2), 1 / log(abs(-2)))))

# "Backward" composition operator composes from right to left, like compose()
f2 <- inv \%<<<\% log \%<<<\% abs
stopifnot(isTRUE(all.equal(f2(-2), f1(-2))))

# Forward composition operator composes from left to right
f3 <- abs \%>>>\% log \%>>>\% inv
stopifnot(isTRUE(all.equal(f3(-2), f1(-2))))

# Argument signature of inner most function is preserved
f <- function(a, b = 0) a + b
stopifnot(identical(formals(compose(inv, f)), formals(f)))

}
