% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry.R
\name{curry}
\alias{curry}
\alias{uncurry}
\title{Curry a function}
\usage{
curry(f)

uncurry(f)
}
\arguments{
\item{f}{Function.}
}
\value{
\code{curry()} returns a curried function.
}
\description{
\code{curry()} \href{https://en.wikipedia.org/wiki/Currying}{curries} functions in a
manner congenial to R’s calling convention.
}
\examples{
f <- function(x, y, z = 3) c(x, y, z)
fc <- curry(f)

stopifnot(
  identical(
    formals(fc(1)),
    formals(function(y, z = 3) {})
  ),
  identical(
    formals(fc(y = 2, z = 4)),
    formals(function(x) {})
  ),
  identical(fc(1)(2),     c(1, 2, 3)),
  identical(fc(y = 2)(1), c(1, 2, 3)),
  identical(fc(1, 2),     c(1, 2, 3)),
  identical(fc(2, x = 1), c(1, 2, 3))
)

double <- curry(`*`)(2)
stopifnot(double(3) == 6)

}
\seealso{
\code{\link[=fn_curry]{fn_curry()}}, \code{\link[=partial]{partial()}}
}
