% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry.R
\name{curry}
\alias{curry}
\alias{curry_fn}
\title{Curry a function}
\usage{
curry(f, env = environment(f))

curry_fn(..., ..env = parent.frame())
}
\arguments{
\item{f}{Function.}

\item{env}{Environment of the curried function or \code{NULL}. If \code{NULL}, the
environment of the curried function is the calling environment.}

\item{...}{Function declaration, which supports
\link[rlang:quasiquotation]{quasiquotation}.}

\item{..env}{Environment in which to create the function (i.e., the
function’s \link[base:environment]{enclosing environment}).}
}
\value{
The original function as nested calls of functions of successive
arguments.
}
\description{
Curry a function
}
\examples{
curry(`*`)
curry(function(x, y, z = 0) x + y + z)
curry_fn(x, y, z = 0 ~ x + y + z)
curry_fn(target, ... ~ identical(target, ...))
compare_to <- curry_fn(target, x ~ identical(QUQ(target), x))
is_this <- compare_to("this")
is_this("that")  # FALSE
is_this("this")  # TRUE

}
