% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{eff}
\alias{eff}
\alias{as_eff}
\title{Low-cost anonymous function}
\usage{
eff(...)

as_eff(x)
}
\arguments{
\item{...}{Function declaration.}
}
\description{
\code{eff()} enables you concisely create \dQuote{anonymous} functions of
arbitrary signature.
}
\examples{
f <- eff(x, y := x + y)
f(1, 2)

f <- eff(x, y = 1 := x + y)
f(1)

f <- eff(x, y = 1, ... := log(x + y, ...))
f(1, 1, base = 2)

f <- eff(x, ... = , y := log(x + y, ...))
f(1, base = 2, y = 1)

zero <- 0
is_positive <- eff(x := x > !! zero)
is_positive(1)
is_positive(-1)

}
