% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff.R
\name{eff}
\alias{eff}
\alias{..}
\alias{λ}
\title{Low-cost anonymous functions}
\usage{
eff(..., ..env = parent.frame())

..(..., ..env = parent.frame())

λ(..., ..env = parent.frame())
}
\arguments{
\item{...}{Function declaration (see below).}

\item{..env}{Environment in which to create the function (i.e., the
function’s enclosing environment).}
}
\value{
A function whose enclosing environment is \code{..env} (by default, the
calling environment of \code{eff()}).
}
\description{
\code{eff()} enables you to concisely create (anonymous) functions of arbitrary
call signature. For even less visual noise, \code{..()} and \code{λ()} are provided as
aliases of \code{eff()}.
}
\section{Function declarations}{
 A \emph{function declaration} is a concise
expression that specifies a function’s arguments and body. It is a
comma-separated expression of the form\preformatted{    arg1, arg2, ..., argN ~ body
}

or\preformatted{    arg1, arg2, ..., argN, ~ body
}

(Note the final, separating comma in the second case.)

To the left of \code{~}, you write a conventional function-argument declaration,
as you’d do in \code{function(...)}: each of \code{arg1}, \code{arg2}, \dots, \code{argN} is
either a bare argument (e.g., \code{x} or \code{...}) or an argument with default
value (e.g., \code{x = 1}). To the right of \code{~}, you write the function body,
i.e., an expression of the arguments.
}

\examples{
f <- eff(x ~ x + 1)
f(1)

f <- eff(x, y ~ x + y)
f(1, 2)

f <- eff(x, y = 2 ~ x + y)
f(1)

f <- eff(x, y = 1, ... ~ log(x + y, ...))
f(1, 1, base = 2)

# to specify '...' in the middle of the call signature, write '... = '
f <- eff(x, ... = , y ~ log(x + y, ...))
f(1, base = 2, y = 1)

# use one-sided formula for constant functions or commands
eff(~ NULL)
eff(~ message("!"))

# unquoting (via `!!` or UQ()) is supported
zero <- 0
eff(x = UQ(zero) ~ x > !! zero)

}
