% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nofrills.R
\docType{package}
\name{nofrills-package}
\alias{nofrills}
\alias{nofrills-package}
\title{nofrills: Low-Cost Anonymous Functions}
\description{
Provides a more concise way to construct functions using the same
syntax as 'function()', with support for Tidyverse-style quasiquotation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egnha/nofrills}
  \item Report bugs at \url{https://github.com/egnha/nofrills/issues}
}

}
\author{
\strong{Maintainer}: Eugene Ha \email{eha@posteo.de}

}
\keyword{internal}
